Inject network spots into NexGen.
This script reads the affiliate log for a day then prepares the required
spots to be loaded into NexGen automation. Expects the necessary media
has been downloaded, but checks in multiple locations. (If a download
was skipped on one channel the same copy may still be found in another
place.)

Had to be rewritten often to keep up with changes in the web sites.